Fixed bug involving selecting by aspect ratio
It was requiring the image to be larger than the largest image in the DB with that aspect ratio rather than the largest in the size set.

Also, I didn't make the connection as to why this was fixed as well, but this was apparently also the reason my images were saving to every size from another size set's sizes of comparable ratio.