Handle exception when creating crop on demand (see desc)
When an Image has been reserved that was originally created before a size was added at a given ratio, on-demand crops will come up with multiple Crop objects from their get() call, causing an exception.  I removed the get, so it can receive multiple results and grab the one with the largest crop_w.