Fix a pre-django 1.6 bug with redundant prefetch_related args
If the following queryset is attempted::

    qset.prefetch_related('image_field', 'image_field__thumbs')

This will raise the error:

    AttributeError: Cannot find 'thumbs' on CropDusterImageFieldFile object,
    'field_name__thumbs' is an invalid parameter to prefetch_related()

In order to have this work in Django 1.4 and 1.5, we need
CropDusterImageFieldFile.thumbs to return a descriptor, which we proxy
from cropduster.models.Image.thumbs. We also need
cropduster_image_fieldfile_instance.thumbs to return the value of
cropduster_image_fieldfile_instance.related_object.thumbs.