Fix bug in best_fit calc where scaling caused img dim to drop below mins
At a certain stage in the best fit calculation it scales down the width
to match the decrease in height due to the initial fit being out of
bounds (or scales down the height to match the decrease in width, as the
case might be). It was not checking that this decrease did not cause the
width or height to drop below the min_w or min_h, respectively.