Fix ReverseForeignRelation regression where limit_choices_to broke
We don't use this functionality in django-cropduster itself, but it
turns out that other parts of The Atlantic codebase do depend on
limit_choices_to being able to filter on the queryset of the reverse
foreign related manager.

This functionality is somewhat orthogonal to django-cropduster and
should perhaps at some point be broken out into its own project.

Also added regression tests for ReverseForeignRelation.