allow re-upload when images not found
This handles the pesky use-case of working with the admin in an environment different
from the one in which an image was initially uploaded with CropdusterField. Prior to
this change, the admin would simply return a 500. After this change, it should instead
allow for users to access the re-upload form.

refs: FUN-2258