Fix upload bug when browser zoom is less than 100%
To reproduce the bug, open chrome and zoom out to some value less than
100%. Then click "Upload Image" to open the cropduster upload
modal. Throttle the connection to "Slow 3G" speed in the Network tab of
the Developer Console and set a breakpoint for the conditional
`if ($('#cropbox').width() > 1) {` in upload.js. Then upload an
image. When the breakpoint is hit, the not-yet-loaded image should have
a width and height greater than 1 (at 90% zoom, for instance, it is
1.11111). Jcrop is then initialized with dimensions of 1.11111x1.11111,
which shows up as a tiny box to the user.